refactor(kernel): dedupe document attestation and auth page session logic
