+++ apps/kernel/app/auth/api/documents/route.ts
@@ -5,10 +5,8 @@ import { corsHeaders } from '@imajin/config';
 import { requireAuth } from '@/src/lib/auth/middleware';
 import { createLogger } from '@imajin/logger';
 import {
-  buildDocumentSignatureRows,
-  getCreatorDisplayName,
+  finalizeDocumentAttestation,
   parseDocumentRequestBody,
-  publishDocumentCreatedNotifications,
   validateDocumentRequestInput,
 } from '../../../../src/lib/auth/document-attestation';
 
@@ -93,37 +91,17 @@ export async function POST(request: NextRequest) {
     })
     .returning();
 
-  // Create signature rows
-  const sigRows = buildDocumentSignatureRows({
-    attestationId,
-    creatorDid: callerDid,
-    creatorJws: authorJws,
-    signerDids,
-    genId,
-  });
-
-  await db.insert(attestationSignatures).values(sigRows);
-
-  // Set asset immutable
-  await db.update(assets).set({ immutable: true }).where(eq(assets.id, documentAssetId));
-
-  // Publish one event per pending signer so notify reactors can target recipients directly.
-  const creatorName = getCreatorDisplayName(callerIdentity, callerDid);
-  publishDocumentCreatedNotifications({
+  const allSigs = await finalizeDocumentAttestation({
     attestationId,
     documentAssetId,
     creatorDid: callerDid,
-    creatorName,
+    creatorJws: authorJws,
     signerDids,
     title,
+    callerIdentity,
+    genId,
     log,
   });
-
-  // Fetch signatures to return
-  const allSigs = await db
-    .select()
-    .from(attestationSignatures)
-    .where(eq(attestationSignatures.attestationId, attestationId));
 
   return NextResponse.json(
     { attestation, signatures: allSigs },
